Refactor axis management, enhance memory handling, and improve command validation
- Updated `create_axis` and `delete_axis` functions for better memory management, including accurate size calculations and efficient data movement.
- Refactored `create_link` and `delete_link` functions to streamline link management, ensuring proper memory adjustments and file synchronization.
- Introduced `ensure_axis_exists` function to validate and create axes as needed, improving command handling in the CLI.
- Enhanced argument validation across command handlers, including `check-axis` and `create-link`, to ensure robust user interaction.
- Updated documentation to reflect new features, including detailed examples of axis and link management operations, size calculations, and memory layout strategies.

+++ src/channel.h
@@ -9,8 +9,8 @@
 
 // Function declarations
 ushort get_channel_count(uchar* node);
-uint get_channel_offset(uchar* node, int channel_index);
-ushort get_channel_size(uchar* node, int channel_index);
-int create_channel(int node_index);
+uint get_channel_offset(uchar* node, ushort channel_index);
+ushort get_channel_size(uchar* node, ushort channel_index);
+int create_channel(uint node_index);
 
 #endif
